requirements.txt
pandas
requests
psycopg2-binary
pyarrow
python-dotenv